chore : 패키지명 변경에 따른 경로 수정
