+++ core/viewmodel/src/commonMain/kotlin/com/whatever/caro/core/viewmodel/contract/UiIntent.kt
@@ -1,0 +1,3 @@
+package com.whatever.caro.core.viewmodel.contract
+
+interface UiIntent
