+++ core/viewmodel/src/commonMain/kotlin/com/whatever/caro/core/viewmodel/contract/UiState.kt
@@ -1,0 +1,3 @@
+package com.whatever.caro.core.viewmodel.contract
+
+interface UiState
