--- core/viewmodel/src/commonMain/kotlin/com/whatever/caro/viewmodel/contract/SideEffect.kt
@@ -1,3 +1,0 @@
-package com.whatever.caro.viewmodel.contract
-
-interface SideEffect
